Add a 404 page.
This is an error page we will show if the user
tries to load a path without a route.

I grabbed the text from the existing edd.ca.gov site.
Note that for now the Spanish translation matches the
English translation. The edd.ca.gov doesn't have a
Spanish 404 page, so at least we're consistent. When
we're getting Spanish strings in the future, we should
include the 404 page strings.