Create routes for the various guide pages.
We declare all the known paths in src/data/routes.js (shared by
both the client and the server). If the path isn't in routes.js,
the server returns a 404. In the client, we use a <Switch> to
load the correct path.

This does change user behavior because it introduces the 404 page.
Previously, we redirected everything to the guide, but now we
show a 404 page except for / and /guide/, which both get redirected
to /guide/benefits.