Creates routes for all the guide pages (#243)
* Add a 404 page.

This is an error page we will show if the user
tries to load a path without a route.

I grabbed the text from the existing edd.ca.gov site.
Note that for now the Spanish translation matches the
English translation. The edd.ca.gov doesn't have a
Spanish 404 page, so at least we're consistent. When
we're getting Spanish strings in the future, we should
include the 404 page strings.

* Add a redirect to guide page.

This page just redirects the user to the guide.

* Create routes for the various guide pages.

We declare all the known paths in src/data/routes.js (shared by
both the client and the server). If the path isn't in routes.js,
the server returns a 404. In the client, we use a <Switch> to
load the correct path.

We need a single route for the guide page, otherwise
we break the routing that happens in the TabbedContainer.
This broke the scrolling to the top of the page after
a link click and it broke the keyboard navigation of
tabs as the page was re-rendered.