Refactor to auth logic into <RetroCertsRoute> (#298)
* Refactor to auth logic into <RetroCertsRoute>

This moves the logic to check if the user is logged in
and refetch data with the auth token into <RetroCertsRoute>.
This means for new pages, when adding to App.js, you can
add requiresAuthentication to the route and it will handle
auth logic for you.

The behavior should be the same as before this change, but
less duplicated code. This also makes testing Page components
easier.