Add the confirmation page.
This checks to see that there is a confirmation number.
If one does not exist, it tries to redirect to the
"What to Expect" page.

When logging in, we route directly to this page if there
is a confirmation number. There's no other route to
this page yet, but it should be added once the user
submits.

Also switch to using the same checkbox (from bootstrap)
for the green alerts.