Connect the API to the staging DB. (#301)
Remove test-accounts.js and connect to the Cosmos DB!
The first two accounts exist in the staging DB so
the same logins should work. Note that if multiple
people use the same login, it will invalidate the
first user's token!

Also remove the last name from the "What to expect"
page and add an error code for invalid recaptcha. That
check takes precedence and will show whether your auth
data is correct or not.

Update the tests so we don't try to connect to the
network except for the single ReCaptcha test.