Add the employer income section. (#321)
If the user answered 'Yes' to question 6,
show the section for adding income sources.
The user can add additional income sources
as needed. All the data gets written back
to the react state in userData.formData.

Still needs validation and actually sending
the data to the backend (instead of the
mock data).