Update logic for reading user data from db (#362)
This PR updates the logic for reading from the db to account for the following:

Most (99.99%) users have the same seekWorkPlan for all weeks in weeksToCertify, in which case their seekWorkPlan is a one element array. The rest have an array whose length matches the weeksToCertify length.

Note that this logic should continue to work with all existing test data created under the prior assumption that all seekWorkPlan arrays would be the same length as the weeksToCertify array for that row.