Fix bug where non-submits were considered complete
The bug was if you had completed the last week to
certify, gone back a week, then gone forward, we
would treat that as the user submitting because
all the necessary data was there.

Instead, add an extra "completed" field for when
the user is on the last page. Only assign a
confirmation number if that field is true.