Add staff view routes (#537)
This PR adds routes for the Staff View. It creates two new routes for the StaffView views that have already been merged:

/retroactive-certification/staff-view
staffViewConfirmation: "/retroactive-certification/staff-view/claimant-info

It redirects the new routes to /retroactive-certification/ on production. A follow-up PR will apply the redirects on staging and/or test based on IP address.